Move util.network.smbios to discovery.hardware, move prtconf to puppet plugin
